migration to sqlite
