+++ front/domotik/package.json
@@ -1,6 +1,6 @@
 {
   "name": "web",
-  "version": "0.2.0",
+  "version": "0.3.0",
   "private": true,
   "scripts": {
     "start": "node ./bin/www",
@@ -12,10 +12,10 @@
     "cookie-parser": "^1.3.5",
     "debug": "^2.2.0",
     "express": "^4.13.3",
-    "influx": "^5.5.1",
     "jade": "^1.11.0",
     "marked": "^0.3.5 ",
     "morgan": "^1.6.1",
-    "serve-favicon": "^2.3.0"
+    "serve-favicon": "^2.3.0",
+    "sqlite3": "^5.0.8"
   }
 }
